third_party/WebKit/LayoutTests/fast/events/simulated-key-state-expected.txt

CONSOLE WARNING: line 27: A DOM event generated from JavaScript has triggered a default action inside the browser. This behavior is non-standard and will be removed in M53, around September 2016. See https://www.chromestatus.com/features/5718803933560832 for more details.
This tests that modifier keys are propagated to the fake mouse event created when you press return and a link has focus.

If the test succeeds, you should see six "PASS" messages below.

This is the link used for testing.

PASS: click event had all the right key state.
PASS: click event had all the right key state.
PASS: click event had all the right key state.
PASS: click event had all the right key state.
PASS: click event had all the right key state.
PASS: click event had all the right key state.
